Add files via upload
Adds methods to NSObject, NSNumber and NSString to support intValue, floatValue, doubleValue, boolValue and stringValue on all three objects uniformly. Allows (for example) pulling an NSObject from a dictionary which may be an NSString or an NSNumber and retrieving it's floatValue without getting an exception or having to attempt to cast to (NSNumber) first.