Extend NSNumber / NSObject / NSString
Adds support for intValue, floatValue, doubleValue, boolValue to NSNumber, NSObject, NSString. this allows these value types to be extracted for dictionary entries without having to type check them first.

Implementation note: NSObject returns 0, false, "" rather than throwing a type exception when the NSObject is not an NSNumber or NSString. This is an implementation choice.