ASCIIPropertyListParser: handle non-7b-ASCII chars
Currently, ASCIIPropertyListParser takes bytes[] and then pads the bytes
with an extra 00 byte to get UTF-16. If the byte is >= 0x80, then it
pads it with 0xff. This means that if the bytes are in the 7-bit ASCII
range, everything is fine. But if not, 0x80 for example becomes 0xff80,
(half-width TA katakana) which I don't believe corresponds to any real
encoding system.

The options are to:
 - convert using the default system encoding
 - convert using UTF-8

I think UTF-8 is a better default. The default system encoding is
good for backwards compatibility, but this feature (non-7-bit ASCII)
has never worked at all before, so that's not really necessary. This can
also be made configurable if the need presents itself.