Occasionally (rarely), an NSString content is a string that ends with a NULL-character. This poses a problem for XML serialization, as it is forbidden by the XML specification for a document to contain a null character. I have updated NSString.toXML to truncate string content to the first occurrence of 0x00.
I also noticed that the utf8Encoder code in the same method was ultimately doing nothing, so I've removed it. Changing the encoding of bytes in a String and using them to create a new String in Java will do nothing -- both Strings will have buffers containing underlying UTF16-endoded byte buffers.  And Java Strings are immutable.  This is why Strings can simply be summed for concatenation, without any concern for the underlying encoding.